+++ extensions/msal-node-extensions/test/broker/NativeBrokerPlugin.spec.ts
@@ -1,6 +1,9 @@
 jest.mock("@azure/msal-node-runtime", () => {
+    const actual = jest.requireActual("@azure/msal-node-runtime");
     return {
+        ...actual,
         msalNodeRuntime: {
+            ...actual.msalNodeRuntime,
             SignInSilentlyAsync: jest.fn(),
             SignInAsync: jest.fn(),
             AcquireTokenSilentlyAsync: jest.fn(),
@@ -19,18 +22,6 @@ jest.mock("@azure/msal-node-runtime", () => {
                 SetPopParams: jest.fn(),
                 SetAdditionalParameter: jest.fn(),
             })),
-        },
-        ErrorStatus: {
-            Unexpected: 0,
-            InteractionRequired: 1,
-            AccountUnusable: 2,
-            NoNetwork: 3,
-            NetworkTemporarilyUnavailable: 4,
-            ServerTemporarilyUnavailable: 5,
-            UserCanceled: 6,
-            AuthorityUntrusted: 7,
-            UserSwitched: 8,
-            AccountNotFound: 9,
         },
     };
 });
@@ -64,6 +55,7 @@ import {
 } from "@azure/msal-common";
 import { randomUUID } from "crypto";
 import { NativeAuthError } from "../../src/error/NativeAuthError";
+import { StringUtils } from "../../src/utils/StringUtils.js";
 import {
     testMsalRuntimeAccount,
     testAccountInfo,
@@ -95,9 +87,20 @@ function createMockAuthResult(
 
 if (process.platform === "win32") {
     describe("NativeBrokerPlugin", () => {
+        const enhancedErrorContext = msalRuntimeExampleError.errorContext
+            ? `${msalRuntimeExampleError.errorContext} (Error Code: ${
+                  msalRuntimeExampleError.errorCode
+              }, Tag: ${StringUtils.tagToString(
+                  msalRuntimeExampleError.errorTag
+              )})`
+            : `(Error Code: ${
+                  msalRuntimeExampleError.errorCode
+              }, Tag: ${StringUtils.tagToString(
+                  msalRuntimeExampleError.errorTag
+              )})`;
         const testNativeAuthError = new NativeAuthError(
             ErrorStatus[msalRuntimeExampleError.errorStatus],
-            msalRuntimeExampleError.errorContext,
+            enhancedErrorContext,
             msalRuntimeExampleError.errorCode,
             msalRuntimeExampleError.errorTag
         );
@@ -678,7 +681,7 @@ if (process.platform === "win32") {
                     scopes: testAuthenticationResult.scopes,
                     correlationId: testCorrelationId,
                     authority: testAuthenticationResult.authority,
-                    redirectUri: TEST_REDIRECTURI,
+                    redirectUri: "",
                 };
 
                 const chooseRedirectUriMock = jest.spyOn(
@@ -1515,7 +1518,7 @@ if (process.platform === "win32") {
                     scopes: testAuthenticationResult.scopes,
                     correlationId: testCorrelationId,
                     authority: testAuthenticationResult.authority,
-                    redirectUri: TEST_REDIRECTURI,
+                    redirectUri: "",
                 };
 
                 const chooseRedirectUriMock = jest.spyOn(
@@ -2293,7 +2296,7 @@ if (process.platform === "win32") {
                 scopes: testAuthenticationResult.scopes,
                 correlationId: testCorrelationId,
                 authority: testAuthenticationResult.authority,
-                redirectUri: TEST_REDIRECTURI,
+                redirectUri: "",
             };
 
             const chooseRedirectUriMock = jest.spyOn(
@@ -2338,7 +2341,7 @@ if (process.platform === "win32") {
                 scopes: testAuthenticationResult.scopes,
                 correlationId: testCorrelationId,
                 authority: testAuthenticationResult.authority,
-                redirectUri: TEST_REDIRECTURI,
+                redirectUri: "",
             };
 
             const chooseRedirectUriMock = jest.spyOn(
@@ -2391,7 +2394,7 @@ if (process.platform === "win32") {
                 scopes: testAuthenticationResult.scopes,
                 correlationId: testCorrelationId,
                 authority: testAuthenticationResult.authority,
-                redirectUri: TEST_REDIRECTURI,
+                redirectUri: "",
             };
 
             const chooseRedirectUriMock = jest.spyOn(
@@ -2436,7 +2439,7 @@ if (process.platform === "win32") {
                 scopes: testAuthenticationResult.scopes,
                 correlationId: testCorrelationId,
                 authority: testAuthenticationResult.authority,
-                redirectUri: TEST_REDIRECTURI,
+                redirectUri: "",
             };
 
             const chooseRedirectUriMock = jest.spyOn(