+++ lib/msal-browser/src/custom_auth/core/network_client/custom_auth_api/types/ApiRequestTypes.ts
@@ -14,6 +14,7 @@ export interface SignInInitiateRequest extends ApiRequestBase {
 export interface SignInChallengeRequest extends ApiRequestBase {
     challenge_type: string;
     continuation_token: string;
+    id?: string;
 }
 
 interface SignInTokenRequestBase extends ApiRequestBase {
@@ -31,7 +32,7 @@ export interface SignInOobTokenRequest extends SignInTokenRequestBase {
 }
 
 export interface SignInContinuationTokenRequest extends SignInTokenRequestBase {
-    username: string;
+    username?: string;
 }
 
 /* Sign-up API request types */
@@ -89,4 +90,22 @@ export interface ResetPasswordSubmitRequest extends ApiRequestBase {
 
 export interface ResetPasswordPollCompletionRequest extends ApiRequestBase {
     continuation_token: string;
+}
+
+/* Register API request types */
+export interface RegisterIntrospectRequest extends ApiRequestBase {
+    continuation_token: string;
+}
+
+export interface RegisterChallengeRequest extends ApiRequestBase {
+    continuation_token: string;
+    challenge_type: string;
+    challenge_target: string;
+    challenge_channel?: string;
+}
+
+export interface RegisterContinueRequest extends ApiRequestBase {
+    continuation_token: string;
+    grant_type: string;
+    oob?: string;
 }